Rename reward model trainer and restore original report
