--- report.txt
@@ -1,63 +1,0 @@
-# Fine-Tuning InstructGPT with PPO in the RLHF Stage: A Beginner’s Guide
-
-## Introduction
-InstructGPT, the precursor to ChatGPT, is trained on large-scale text corpora and further aligned to human preferences through reinforcement learning from human feedback (RLHF). During the RLHF stage we replace the next-token prediction objective with a reward signal learned from human preference data. This guide walks through how Proximal Policy Optimization (PPO) is used to fine-tune the model, how standard RL concepts map to language modeling, and how the reward model and PPO loss function interact during training.
-
-## RLHF Basics for Language Models: State, Action, and Reward
-- **State** – the prompt or conversation context presented to the model. As the model generates a response token by token, the partial output can be viewed as the evolving state at each step.
-- **Action** – the token (or sequence of tokens) produced by the model in response to the current state. A complete answer corresponds to a sequence of such actions.
-- **Reward** – a scalar score derived from human judgments. Annotators rank multiple responses to the same prompt and a reward model is trained on those comparisons. The reward model then scores new prompt/response pairs during PPO training.
-
-Think of the model as a student answering questions: the prompt is the question, the response is the student’s answer, and the reward is the teacher’s grade. Over many iterations the model learns which styles of answers consistently earn higher scores.
-
-## What Is PPO and Why Use It in RLHF?
-PPO is a policy-gradient algorithm tailored for stability when updating large neural policies. It is used in RLHF because it:
-- **Limits update size** through clipping or KL penalties so the policy cannot drift abruptly, improving stability.
-- **Encourages exploration** via an entropy bonus, preventing the model from collapsing to deterministic responses too early.
-- **Reuses data efficiently**, enabling multiple optimization steps on the same batch of trajectories—a crucial property when human feedback data is expensive to collect.
-- **Remains simple to implement** compared with alternatives such as TRPO while retaining many of the same trust-region benefits.
-
-PPO therefore strikes a balance between making meaningful progress on the reward objective and avoiding destructive updates that would destabilize the language model.
-
-## PPO in Action: Fine-Tuning InstructGPT with Human Feedback
-1. **Initialize from the supervised model** – start PPO from the supervised fine-tuned (SFT) checkpoint and keep a frozen reference copy for KL regularization.
-2. **Sample a prompt** – prompts are drawn independently, so each PPO episode is effectively a one-step bandit problem.
-3. **Generate a response** – the policy rolls out a response token by token to form a complete answer.
-4. **Score with the reward model** – the frozen reward model evaluates the prompt/response pair and produces a scalar reward calibrated so the SFT baseline averages near zero.
-5. **Estimate advantages** – compare the observed reward against value-function predictions to compute advantages that guide gradient updates.
-6. **Update the policy and value head** – apply PPO’s clipped objective while incorporating the KL penalty and entropy bonus. Multiple gradient steps may be taken per batch.
-
-### Turning Human Preferences into a Reward Signal
-- Collect paired comparisons of model answers from human labelers.
-- Train the reward model (often a scalar head on top of a GPT body) with a Bradley–Terry objective so preferred answers receive higher scores.
-- Freeze the reward model during PPO so the definition of “good” remains consistent across policy updates.
-- Monitor for reward hacking by balancing the PPO objective with a KL penalty that keeps the policy near the reference model.
-
-### Analogy
-If the model is a student, the reward model is an automated grader distilled from many teacher evaluations. PPO tells the student to adjust their approach gradually—small improvements at every assignment build better alignment without wild swings in behavior.
-
-## How the Reward Signal Updates the Model
-- **Policy gradients** weight the log-probability of generated tokens by the computed advantages, reinforcing high-reward actions and suppressing low-reward ones.
-- **Token-level credit assignment** allows PPO to identify which parts of a response contributed to the final reward, adjusting probabilities for specific phrases or behaviors.
-- **Baseline subtraction** via the value function reduces variance so learning focuses on unexpected outcomes.
-- **KL regularization** ensures the policy does not exploit quirks of the reward model by straying too far from the reference distribution.
-- **Iterative refinement** gradually steers the model toward answers that are more helpful, harmless, and honest according to the learned reward.
-
-## Conclusion
-RLHF with PPO equips language models to internalize human preferences without needing explicit ground-truth answers for every prompt. The policy treats prompts as states, responses as actions, and human-derived rewards as feedback. PPO’s clipped updates, advantage estimation, entropy bonus, and KL penalty make training stable while still allowing the policy to improve. Over many iterations the policy converges toward responses that humans consistently prefer, leading to aligned systems like InstructGPT and ChatGPT.
-
-## References
-- OpenAI, *Training language models to follow instructions with human feedback* (InstructGPT).
-- Schulman et al., *Proximal Policy Optimization Algorithms*.
-- Blog posts and articles summarizing RLHF best practices.
-
-## Coding Assignment
-Create a compact RLHF training pipeline for GPT-2 Small using only PyTorch:
-
-1. **Project structure** – separate supervised fine-tuning, reward-model training, and PPO policy optimization modules.
-2. **Model components** – implement GPT backbones, scalar reward/value heads, and data utilities for prompt/response preferences.
-3. **Training routines** – write functions for SFT, reward model fitting, and PPO with configurable hyperparameters and logging hooks.
-4. **Unit tests** – include gradient checks for the SFT loss and PPO objective to verify correctness of analytic gradients.
-5. **Implementation notes** – stub out function bodies with clear docstrings describing the mathematics each implementation must follow.
-
-The end goal is a working RLHF pipeline driven entirely by PyTorch, ready for students to complete by filling in the provided skeleton code and passing the gradient-check unit tests.
