+++ tests/test_ppo.py
@@ -9,7 +9,7 @@ ROOT = os.path.dirname(os.path.dirname(__file__))
 sys.path.append(ROOT)
 
 from gpt import GPTConfig, GPT
-from reward_model import ScalarHead
+from train_rm import ScalarHead
 from train_ppo import PPOTrainer, gather_log_probs
 from train_ppo import train_ppo  # noqa: F401  # ensure module imports without argparse
